# contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Handdraw API",
    "version": "1.0.0",
    "description": "Handdraw API contract. Identity, workspace onboarding and project/board routes are wired behind explicit configuration. Board deletion additionally requires the cleanup worker. Team membership, invitations and board Viewer grants require the membership feature flag. Invitation email delivery remains pending integration. Bearer identity is verified before profile resolution. Unknown fields/trailing JSON are rejected. Future domain endpoints are added with their implementation contracts."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8081",
      "description": "Local development; deployment must use HTTPS."
    }
  ],
  "security": [
    {
      "supabaseBearer": []
    }
  ],
...
  "paths": {
...
    "/v1/workspaces/{workspace_id}/members": {
      "get": {
        "operationId": "listMembers",
        "summary": "List members",
        "parameters": [
          {
            "$ref": "#/components/parameters/workspace_id"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "x-max-request-bytes": 65536,
        "tags": [
          "Members"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available with APP_MEMBERSHIP_ENABLED, board/workspace/identity enabled and schema version 9. Available when the corresponding runtime feature is enabled. List the workspace roster without exposing Auth email addresses. Cursor ordering: resource ID ascending."
      }
    },
    "/v1/workspaces/{workspace_id}/members/{user_id}": {
      "patch": {
        "operationId": "changeMemberRole",
        "summary": "Change member role",
        "parameters": [
          {
            "$ref": "#/components/parameters/workspace_id"
          },
          {
            "$ref": "#/components/parameters/user_id"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "pattern": "^\"[1-9][0-9]*\"$"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "422": {
            "$ref": "#/components/responses/422"
          },
          "412": {
            "$ref": "#/components/responses/412"
          },
          "428": {
            "$ref": "#/components/responses/428"
          }
        },
        "x-max-request-bytes": 65536,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeMember"
              }
            }
          }
        },
        "tags": [
          "Members"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available with APP_MEMBERSHIP_ENABLED, board/workspace/identity enabled and schema version 9. Available when the corresponding runtime feature is enabled. Owner changes Editor/Viewer role using the current member revision and atomic seat capacity checks."
      },
      "delete": {
        "operationId": "removeMember",
        "summary": "Remove member",
        "parameters": [
          {
            "$ref": "#/components/parameters/workspace_id"
          },
          {
            "$ref": "#/components/parameters/user_id"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "responses": {
          "204": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "422": {
            "$ref": "#/components/responses/422"
          },
          "412": {
            "$ref": "#/components/responses/412"
          },
          "428": {
            "$ref": "#/components/responses/428"
          }
        },
        "x-max-request-bytes": 65536,
        "tags": [
          "Members"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available with APP_MEMBERSHIP_ENABLED, board/workspace/identity enabled and schema version 9. Available when the corresponding runtime feature is enabled. Remove a non-Owner member; an authorized repeat has no further effect."
      }
    },
    "/v1/workspaces/{workspace_id}/invitations": {
      "post": {
        "operationId": "inviteMember",
        "summary": "Invite member",
        "parameters": [
          {
            "$ref": "#/components/parameters/workspace_id"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "202": {
            "description": "Invitation persisted; email delivery remains pending integration. Concurrent processing may return PendingOperationResponse.",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InvitationSubmissionResponse"
                    },
                    {
                      "$ref": "#/components/schemas/PendingOperationResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "x-max-request-bytes": 65536,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Invite"
              }
            }
          }
        },
        "tags": [
          "Invitations"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available with APP_MEMBERSHIP_ENABLED, board/workspace/identity enabled and schema version 9. Available when the corresponding runtime feature is enabled. Owner reserves Team Editor capacity and records a seven-day recipient-bound invitation; email delivery is pending integration."
      },
      "get": {
        "operationId": "listInvitations",
        "summary": "List invitations",
        "parameters": [
          {
            "$ref": "#/components/parameters/workspace_id"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "x-max-request-bytes": 65536,
        "tags": [
          "Invitations"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available with APP_MEMBERSHIP_ENABLED, board/workspace/identity enabled and schema version 9. Available when the corresponding runtime feature is enabled. Owner lists invitation history, including calculated expiry. Cursor ordering: created_at descending, then invitation ID descending."
      }
    },
    "/v1/boards/{board_id}/invitations": {
      "post": {
        "operationId": "inviteBoardGuest",
        "summary": "Invite board guest",
        "parameters": [
          {
            "$ref": "#/components/parameters/board_id"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "202": {
            "description": "Invitation persisted; email delivery remains pending integration. Concurrent processing may return PendingOperationResponse.",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InvitationSubmissionResponse"
                    },
                    {
                      "$ref": "#/components/schemas/PendingOperationResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "x-max-request-bytes": 65536,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteGuest"
              }
            }
          }
        },
        "tags": [
          "Invitations"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available with APP_MEMBERSHIP_ENABLED, board/workspace/identity enabled and schema version 9. Available when the corresponding runtime feature is enabled. Owner records a Viewer-only invitation for one Cloud/Team board; email delivery is pending integration."
      }
    },
    "/v1/invitations/accept": {
      "post": {
        "operationId": "acceptInvitation",
        "summary": "Accept invitation",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedAccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "422": {
            "$ref": "#/components/responses/422"
          },
          "410": {
            "$ref": "#/components/responses/410"
          }
        },
        "x-max-request-bytes": 65536,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitation"
              }
            }
          }
        },
        "tags": [
          "Invitations"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available with APP_MEMBERSHIP_ENABLED, board/workspace/identity enabled and schema version 9. Available when the corresponding runtime feature is enabled. Consume one token after matching the current confirmed Auth email and rechecking entitlement/capacity. Repeat use returns 410; inspect current access after a lost response."
      }
    },
    "/v1/invitations/{invitation_id}": {
      "delete": {
        "operationId": "revokeInvitation",
        "summary": "Revoke invitation",
        "parameters": [
          {
            "$ref": "#/components/parameters/invitation_id"
          }
        ],
        "responses": {
          "204": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "x-max-request-bytes": 65536,
        "tags": [
          "Invitations"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available with APP_MEMBERSHIP_ENABLED, board/workspace/identity enabled and schema version 9. Available when the corresponding runtime feature is enabled. Owner revokes a pending invitation and releases reserved capacity."
      }
    },
    "/v1/boards/{board_id}/guests": {
      "get": {
        "operationId": "listBoardGuests",
        "summary": "List board guests",
        "parameters": [
          {
            "$ref": "#/components/parameters/board_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "x-max-request-bytes": 65536,
        "tags": [
          "Guests"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available with APP_MEMBERSHIP_ENABLED, board/workspace/identity enabled and schema version 9. Available when the corresponding runtime feature is enabled. Owner lists explicit Viewer grants, independently of workspace membership."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "supabaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Supabase JWT"
      }
    },
    "schemas": {
...
      "MemberListResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "meta": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "request_id": {
                "type": "string",
                "minLength": 1
              },
              "pagination": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "next_cursor": {
                    "type": "string",
                    "maxLength": 4096,
                    "nullable": true
                  }
                },
                "required": [
                  "next_cursor"
                ]
              }
            },
            "required": [
              "request_id",
              "pagination"
            ]
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Member"
            }
          }
        },
        "required": [
          "success",
          "meta",
          "result"
        ]
      },
...
    },
    "parameters": {
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 4096
        },
        "description": "Opaque authenticated cursor bound to actor, workspace, resource, filter and the operation-specific ordering described below; mismatch is invalid_cursor."
      },
...
      "workspace_id": {
        "name": "workspace_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/WorkspaceID"
        }
      },
...
      "board_id": {
        "name": "board_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/BoardID"
        }
      },
      "user_id": {
        "name": "user_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/UserID"
        }
      },
...
    },
...
  }
}